+++ R/aa_helper_functions.R
@@ -879,13 +879,16 @@ font_green_bg <- function(..., collapse = " ") {
   try_colour(..., before = "\033[42m", after = "\033[49m", collapse = collapse)
 }
 font_rsi_R_bg <- function(..., collapse = " ") {
-  try_colour(..., before = "\033[48;5;210m", after = "\033[49m", collapse = collapse)
+  #ED553B
+  try_colour(..., before = "\033[48;5;203m", after = "\033[49m", collapse = collapse)
 }
 font_rsi_S_bg <- function(..., collapse = " ") {
-  try_colour(..., before = "\033[48;5;113m", after = "\033[49m", collapse = collapse)
+  #3CAEA3
+  try_colour(..., before = "\033[48;5;79m", after = "\033[49m", collapse = collapse)
 }
 font_rsi_I_bg <- function(..., collapse = " ") {
-  try_colour(..., before = "\033[48;5;185m", after = "\033[49m", collapse = collapse)
+  #F6D55C
+  try_colour(..., before = "\033[48;5;222m", after = "\033[49m", collapse = collapse)
 }
 font_red_bg <- function(..., collapse = " ") {
   try_colour(..., before = "\033[41m", after = "\033[49m", collapse = collapse)