+++ tests/testthat/test-mic.R
@@ -51,6 +51,13 @@ test_that("mic works", {
   pdf(NULL) # prevent Rplots.pdf being created
   expect_silent(barplot(as.mic(c(1, 2, 4, 8))))
   expect_silent(plot(as.mic(c(1, 2, 4, 8))))
+  expect_silent(plot(as.mic(c(1, 2, 4, 8)), expand = FALSE))
+  expect_silent(plot(as.mic(c(1, 2, 4, 8)), mo = "esco", ab = "cipr"))
+  if (require("ggplot2")) {
+    expect_s3_class(ggplot(as.mic(c(1, 2, 4, 8))), "gg")
+    expect_s3_class(ggplot(as.mic(c(1, 2, 4, 8)), expand = FALSE), "gg")
+    expect_s3_class(ggplot(as.mic(c(1, 2, 4, 8, 32)), mo = "esco", ab = "cipr"), "gg")
+  }
   expect_output(print(as.mic(c(1, 2, 4, 8))))
   
   expect_equal(summary(as.mic(c(2, 8))),